order_service/app/middlewares/auth_middleware.py
import jwt
from fastapi import HTTPException, Header, Depends
from app.config.settings import SECRET_KEY, ALGORITHM

def get_current_user(authorization: str = Header(None)):
    """Vérifie l'authentification via JWT et retourne l'utilisateur"""
    if not authorization or not authorization.startswith("Bearer "):
        raise HTTPException(status_code=401, detail="Missing Authorization Header")
    
    token = authorization.split(" ")[1]
    
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload  # Retourne les infos du token
    except jwt.ExpiredSignatureError:
        raise HTTPException(status_code=401, detail="Token expired")
    except jwt.InvalidTokenError:
        raise HTTPException(status_code=401, detail="Invalid token")

def check_role(allowed_roles: list):
    """Vérifie si l'utilisateur a le bon rôle"""
    def role_checker(user: dict = Depends(get_current_user)):
        if user["role"] not in allowed_roles:
            raise HTTPException(status_code=403, detail="Access denied")
        return user
    return role_checker
