order_service/requirements.txt
fastapi
uvicorn
pydantic[email]
bcrypt
python-dotenv
python-jose[cryptography]
